Massive restructure for 2022 semester
- Moving all demos into one folder.
- Removing doubled-up resources
- Removing remainders of Daydream code
- Integrating Resonance plugin and Cardboard XR package
- Syncing to latest SentienceLab code
- Extension of Teleporter related scripts

+++ Assets/ResonanceAudio/Shaders/ResonanceAudioSurfaceMaterial.shader
@@ -58,7 +58,7 @@ Shader "ResonanceAudio/SurfaceMaterial" {
         // material of the sub-mesh.
         int subMeshIndex = 0;
         for (; subMeshIndex < _NumSubMeshes; ++subMeshIndex) {
-          if (triangleId < _SubMeshEnds[subMeshIndex]) {
+          if ((float) triangleId < _SubMeshEnds[subMeshIndex]) {
             break;
           }
         }
